pyproject.toml
[project]
name = "first-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "bs4>=0.0.2",
    "fastapi>=0.124.0",
    "mcp[cli]>=1.23.1",
    "python-dotenv>=1.2.1",
]

